+++ src/main/java/edu/colostate/cs/cs414/engima/controllers/Manager.java
@@ -1,4 +1,4 @@
-package edu.colostate.cs.cs414.enigma.users;
+package edu.colostate.cs.cs414.engima.controllers;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
